Fix setColor Int overflow: 0xFF000000.toInt()

+++ app/src/main/java/com/notify/AppMonitorService.kt
@@ -376,7 +376,7 @@ class AppMonitorService : Service() {
         )
         return NotificationCompat.Builder(this, CHANNEL_ID)
             .setSmallIcon(R.drawable.ic_stat_notify)
-            .setColor(0xFF000000)
+            .setColor(0xFF000000.toInt())
             .setLargeIcon(android.graphics.BitmapFactory.decodeResource(resources, AppAlias.iconRes(this)))
             .setContentTitle(AppAlias.labelFor(AppAlias.current(this)))
             .setContentText(statusText())
